added a delete tweet functionality if tweet is mine

+++ index.js
@@ -1,5 +1,5 @@
 import { tweetsData as initialTweetsData } from './data.js'
-import { v4 as uuidv4} from 'https://jspm.dev/uuid'
+import { v4 as uuidv4 } from 'https://jspm.dev/uuid'
 
 let tweetsData = JSON.parse(localStorage.getItem("tweetsData")) || initialTweetsData;
 
@@ -14,14 +14,18 @@ document.addEventListener('click', function (e) {
         let tweetId = e.target.dataset.retweet
         handleRetweetClick(tweetId)
     }
-    if (e.target.dataset.reply){
+    if (e.target.dataset.reply) {
         let tweetId = e.target.dataset.reply
         handleReplyClick(tweetId)
     }
-    if (e.target.id == "tweet-btn"){
+    if (e.target.dataset.delete) {
+        let tweetId = e.target.dataset.delete
+        handleDeleteBtnClick(tweetId)
+    }
+    if (e.target.id == "tweet-btn") {
         handleTweetBtnClick()
     }
-    if (e.target.classList.contains('myreply-button')){
+    if (e.target.classList.contains('myreply-button')) {
         let tweetId = e.target.dataset.tweetid
         handleMyReplyBtnClick(tweetId)
     }
@@ -70,39 +74,39 @@ function handleRetweetClick(tweetId) {
 }
 
 
-function handleReplyClick(replyId){
+function handleReplyClick(replyId) {
     document.getElementById(`replies-${replyId}`).classList.toggle("hidden")
 }
 
 
-function handleMyReplyBtnClick(tweetId){
-    //get the reply textarea for the specific tweet
+function handleMyReplyBtnClick(tweetId) {
+    // get the reply textarea for the specific tweet
     const replyInput = document.getElementById(`reply-input-${tweetId}`)
-    
-    if (!replyInput.value){
+
+    if (!replyInput.value) {
         console.log("you must enter a valid reply")
         return
     }
-    
-    //find the tweet (full xrisimi praktiki)
+
+    // find the tweet (full xrisimi praktiki)
     const targetTweet = tweetsData.find(tweet => tweet.uuid === tweetId)
-    
+
     if (targetTweet) {
-        //create the new reply object
+        // create the new reply object
         const newReply = {
             handle: `@OGTwink`,
             profilePic: `images/literallyme.jpeg`,
             tweetText: replyInput.value,
             uuid: uuidv4()
         }
-        
+
         //add the reply to the target tweet's replies array
         targetTweet.replies.unshift(newReply)
-        
+
         replyInput.value = ''
 
         localStorage.setItem("tweetsData", JSON.stringify(tweetsData))
-        
+
         render()
     }
 }
@@ -110,37 +114,62 @@ function handleMyReplyBtnClick(tweetId){
 
 
 
-//handle new tweets manually inputed by user
-function handleTweetBtnClick(){
-    
+// handle new tweets manually inputed by user
+function handleTweetBtnClick() {
+
     const tweetInput = document.getElementById(`tweet-input`)
 
-    if (!tweetInput.value){
+    if (!tweetInput.value) {
         console.log("you must enter a valid string")
-    } else{
+    } else {
         let newTweet = {
-        handle: `@OGTwink`,
-        profilePic: `images/literallyme.jpeg`,
-        likes: 0,
-        retweets: 0,
-        tweetText: `${tweetInput.value}`,
-        replies: [],
-        isLiked: false,
-        isRetweeted: false,
-        uuid: uuidv4(),
+            handle: `@OGTwink`,
+            profilePic: `images/literallyme.jpeg`,
+            likes: 0,
+            retweets: 0,
+            tweetText: `${tweetInput.value}`,
+            replies: [],
+            isLiked: false,
+            isRetweeted: false,
+            uuid: uuidv4(),
+        }
+
+
+
+        //push  them into the array and make them appear in the screen
+        tweetsData.unshift(newTweet)
+        tweetInput.value = ''
+
+        localStorage.setItem("tweetsData", JSON.stringify(tweetsData))
+
+        render()
     }
 
-    
-    
-    //push  them into the array and make them appear in the screen
-    tweetsData.unshift(newTweet)
-    tweetInput.value = ''
+}
 
-    localStorage.setItem("tweetsData", JSON.stringify(tweetsData))
-    
-    render()
+
+
+// handle the deletion of ones tweet
+function handleDeleteBtnClick(tweetId) {
+
+    const targetTweet = tweetsData.find(tweet => tweet.uuid === tweetId)
+    console.log('targettweet is: ', targetTweet)
+
+    // check if the tweet is mine
+    if (!targetTweet || targetTweet.handle !== "@OGTwink"){
+        console.log('this tweet is not yours')
+        return
     }
-    
+
+    if (confirm('Are you sure you want to delete this twink?')){
+
+        // delete the selected tweet by filtering it out of the tweetsData array 
+        console.log('this tweet is yours. deleting...')
+        tweetsData = tweetsData.filter(tweet => tweet.uuid !== tweetId);
+        localStorage.setItem("tweetsData", JSON.stringify(tweetsData));
+
+        render()
+    }  
 }
 
 
@@ -148,9 +177,9 @@ function handleTweetBtnClick(){
 
 function getFeedHtml() {
     let feedHtml = ``
-        
-    
-    
+
+
+
     tweetsData.forEach(function (tweet) {
 
         //conditional css class switches for likes and retweets
@@ -164,21 +193,21 @@ function getFeedHtml() {
         }
 
 
-        
-        
+
+
         //handle tweet replies
         let repliesHTML = ""
-    
+
         let myReply = `
                 <div class="myreply" id="myreply">
-                    <img src="images/literallyme.jpeg" class="profile-pic">
+                    <img src="images/literallyme.jpeg" class="profile-pic" id="reply-profile-pic">
                     <textarea placeholder="ekfrasou elefthera omorfopaido" class="reply-input" id="reply-input-${tweet.uuid}"></textarea>
                     <button class="myreply-button" id="myreply-button" data-tweetid="${tweet.uuid}">Twink</button>
                 </div>`
 
         if (tweet.replies.length > 0) {
 
-            tweet.replies.forEach(function(reply){
+            tweet.replies.forEach(function (reply) {
                 repliesHTML += `
                 <div class="tweet-reply">
                     <div class="tweet-inner">
@@ -191,11 +220,14 @@ function getFeedHtml() {
                 </div>
                 `
             })
-
-
         }
 
 
+        // check if the reply is made by me and insert it down at the html
+        const deleteButtonHtml = tweet.handle === "@OGTwink" 
+        ? `<i class="delete-tweet" data-delete="${tweet.uuid}">X</i>`
+        : '';
+        
         
         feedHtml +=
             `
@@ -204,6 +236,9 @@ function getFeedHtml() {
                     <img src="${tweet.profilePic}" class="profile-pic">
                     <div>
                         <p class="handle">${tweet.handle}</p>
+                        
+                        ${deleteButtonHtml}
+                        
                         <p class="tweet-text">${tweet.tweetText}</p>
                         <div class="tweet-details">
                             <span class="tweet-detail">
